Added scope check logic
This commit is small but it is VERY important. The main feature I've added here is new dependence: scope check dependence. Logic of this is simple, if you put it on some endpoint - ONLY user with admin scope can use it. It is usefull in different cases, but here i will use mainly in endpoints such as creating, deleting or updating new picture or cologne (in future updates), because only admin can do it. Also in this commit i moved TokenManager class from utils dir into services